Added Score and Reset Functionality
I added a @State var for storing the score and a counter for keeping track of how many times questions have asked. The score was then implemented to the tappedFlag function and output onto the screen and alert. I also made a new function that reset the game after the questions had been asked 8 times and implemented a new alert that showed you the final score out of 8 and allows you to reset the game with a button that uses the reset function.